chore(message): move subcomponents into an internal directory
moves `MessageContent` and `TypeIcon` subcompoentns into an internal directory.

BREAKING CHANGE: `MessageContent` and `TypeIcon` subcomponents have been moved to an `__internal__`
directory
