+++ src/components/message/__internal__/message-content/message-content.component.tsx
@@ -1,16 +1,12 @@
 import React from "react";
 
-import Content from "../../content";
+import Content from "../../../content";
 import MessageContentStyle from "./message-content.style";
 
 export interface MessageContentProps {
-  /** set message title */
   title?: React.ReactNode;
-  /** set content to component */
   children?: React.ReactNode;
-  /** determines if the close icon is shown */
   showCloseIcon?: boolean;
-  /** @ignore @private used to reduce the left-padding for the transparent variant */
   reduceLeftPadding?: boolean;
 }
 
@@ -18,7 +14,7 @@ const MessageContent = ({
   title,
   children,
   showCloseIcon,
-  reduceLeftPadding = false,
+  reduceLeftPadding,
 }: MessageContentProps) => {
   return (
     <MessageContentStyle
