+++ src/components/message/__internal__/message-content/message-content.style.ts
