+++ src/components/message/__internal__/type-icon/type-icon.spec.tsx
@@ -1,24 +1,15 @@
 import React from "react";
 import TestRenderer from "react-test-renderer";
-import { shallow } from "enzyme";
 
 import TypeIcon, { TypeIconProps } from "./type-icon.component";
-import TypeIconStyle from "./type-icon.style";
 
 function render(props?: TypeIconProps) {
-  return TestRenderer.create(<TypeIcon {...props} />);
+  return TestRenderer.create(<TypeIcon variant="info" {...props} />);
 }
 
 const messages = ["info", "error", "success", "warning", "neutral"] as const;
 
 describe("TypeIcon", () => {
-  describe("default props", () => {
-    it("should render with proper default props", () => {
-      const wrapper = shallow(<TypeIcon />);
-      expect(wrapper.find(TypeIconStyle).prop("variant")).toBe("info");
-      expect(wrapper.find(TypeIconStyle).prop("transparent")).toBe(false);
-    });
-  });
   describe("when rendered", () => {
     describe.each(messages)("with no additional props", (variant) => {
       it(`should match the snapshot for ${variant}`, () => {
