+++ src/components/toast/toast.style.ts
@@ -1,5 +1,5 @@
 import styled, { css } from "styled-components";
-import TypeIcon from "../message/type-icon/type-icon.style";
+import TypeIcon from "../message/__internal__/type-icon/type-icon.style";
 import StyledIconButton from "../icon-button/icon-button.style";
 import Portal from "../portal/portal";
 import baseTheme from "../../style/themes/base";
